+++ btclib/utils.py
@@ -41,6 +41,21 @@ from typing import Any
 from btclib.alias import BinaryData, Integer, Octets
 from btclib.exceptions import BTClibTypeError, BTClibValueError
 
+__all__ = [
+    "NoneOneOrMoreInt",
+    "assert_no_trailing",
+    "bytes_from_octets",
+    "bytesio_from_binarydata",
+    "decode_num",
+    "encode_num",
+    "hex_string",
+    "int_from_bits",
+    "int_from_integer",
+    "int_from_json_number",
+    "is_integer",
+    "read_exactly",
+]
+
 NoneOneOrMoreInt = int | Iterable[int] | None